Don't always unlink files in Vault when backup was removed

+++ src/sync.cpp
@@ -3652,13 +3652,13 @@ void Syncs::disableSelectedSyncs(std::function<bool(SyncConfig&, Sync*)> selecto
     if (completion) completion(nDisabled);
 }
 
-void Syncs::removeSelectedSyncs(std::function<bool(SyncConfig&, Sync*)> selector, handle bkpDest)
+void Syncs::removeSelectedSyncs(std::function<bool(SyncConfig&, Sync*)> selector, handle bkpDest, bool ignoreUndefBackupId)
 {
     for (auto i = mSyncVec.size(); i--; )
     {
         if (selector(mSyncVec[i]->mConfig, mSyncVec[i]->mSync.get()))
         {
-            removeSyncByIndex(i, bkpDest);
+            removeSyncByIndex(i, bkpDest, ignoreUndefBackupId);
         }
     }
 }
@@ -3679,7 +3679,7 @@ void Syncs::purgeSyncs()
     if (!mSyncConfigStore) return;
 
     // Remove all syncs.
-    removeSelectedSyncs([](SyncConfig&, Sync*) { return true; });
+    removeSelectedSyncs([](SyncConfig&, Sync*) { return true; }, UNDEF, true);
 
     // Truncate internal sync config database.
     mSyncConfigStore->write(LocalPath(), SyncConfigVector());
@@ -3696,7 +3696,7 @@ void Syncs::purgeSyncs()
     }
 }
 
-void Syncs::removeSyncByIndex(size_t index, handle bkpDest)
+void Syncs::removeSyncByIndex(size_t index, handle bkpDest, bool ignoreUndefBackupId)
 {
     if (index < mSyncVec.size())
     {
@@ -3721,7 +3721,7 @@ void Syncs::removeSyncByIndex(size_t index, handle bkpDest)
             Node* remoteNode = mClient.nodeByHandle(config.getRemoteNode());
             assert(remoteNode && remoteNode->firstancestor()->nodeHandle() == mClient.rootnodes.vault);
 
-            if (bkpDest == UNDEF) // permanently delete
+            if (bkpDest == UNDEF && !ignoreUndefBackupId) // permanently delete
             {
                 mClient.unlink(remoteNode, false, mClient.nextreqtag(), nullptr, true);
             }
